+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2]
+
+## Added
+- to_df and to_fasta: versatile functions 
+- example.py
+- example.fasta
+
+## Changes
+- Better documentation
+- Simpler README
+
 ## [0.0.1]
 
 ### Added